📬 Improve attribute parsing and parse sender report packets (#4)
This change makes some small improvements to attribute parsing during ingest. Audio/video codec, ssrc, payload type, etc. are actually parsed and communicated to the FtlStream to process against incoming media packets instead of hard-coded values. Some rudimentary validation is also performed when the client attempts to begin streaming.

A change was also made to the SDP payload to resolve an issue with Firefox failing to decode the video stream.