Handle pending sessions gracefully (#7)
This change allows for sessions to complete a watch request for a channel ID that does not yet have a stream started. These sessions will be put into a 'pending' state until a stream is available with the given channel ID.

Sessions that exist when a stream has ended will also be put into this 'pending' state until the session ends or the channel ID starts a new stream.

This change also includes a fix for #5.