Update license to AGPLv3 (#34)
This change updates the license of the project from GPLv3 to AGPLv3.

As this software is primarily a network service, it's important to ensure that the open source nature of the project is preserved even if "distribution" occurs via communication to the software over a network server rather than distribution of the software binaries. The AGPL includes some additional provisions over the GPL that allow us to ensure this.