⌛ Fix deadlock on FtlControlConnection::stopConnection (#76)
Fixes #74 

Before this fix, `FtlControlConnection::stopConnection(...)` would call `Stop()` on its underlying `NetworkSocketConnectionTransport`, which in turn would wait for its underlying read thread to exit. Problem is, the original call to `stopConnection(...)` was triggered by the `NetworkSocketConnectionTransport` read thread, resulting in a deadlock as the thread waits for itself to exit.

This change adds a `noBlock` parameter to `ConnectionTransport::Stop(...)`, allowing callers to indicate that it should not block on the exiting of the transport read thread. `FtlControlConnection::stopConnection(...)` uses this parameter to avoid the deadlock.

In the future it might make sense to add a separate "dispatch" thread to `NetworkSocketConnectionTransport` so callbacks aren't tied to the actual read/write thread, which would avoid deadlocks like this one at the cost of a little synchronization overhead.