💀 Fix deadlock on stopping bandwidth capped streams (#85)
Fixes #84

`JanusFtl::serviceReportThreadBody(...)` holds a lock on `JanusFtl::streamDataMutex` while evaluating if clients have exceeded the maximum bitrate setting.

A deadlock can occur if a client exceeding the bitrate cap is stopped while it is in the midst of relaying an RTP packet. Here's the outline:

1. `JanusFtl::serviceThreadReportBody` holds lock on `JanusFtl::streamDataMutex`, calls `FtlServer::StopStream(...)`
2. `FtlServer::StopStream(...)` -> `FtlStream::Stop()` -> `FtlStream::mediaTransport::Stop()`
3. `FtlStream::mediaTransport::Stop()` closes the socket and waits for the connection thread to end.
4. The connection thread is busy relaying an RTP packet in `JanusFtl::ftlServerRtpPacket(...)`, and is attempting to acquire a shared lock on `JanusFtl::streamDataMutex`, but the lock is already being held by the `JanusFtl::serviceThreadReportBody` thread.

This change stops the offending streams without holding a lock on `streamDataMutex`, preventing the deadlock from occurring.