0️⃣ Allow zero length UDP socket reads (#86)
While a zero-length read on a non-blocking TCP socket indicates the client has closed the connection, it does not mean the same for UDP, where zero-length packets are allowed.

Sometimes the FTL client will choke up a bit if the machine is under load and send a few empty UDP packets - we don't want to close their connection in this case. Discovered this while load testing.