🔙 Improved NACKs (#135)
This is a re-write of the NACK feature to make it testable and eliminate edge cases where it would break streams. Shoutout to Hayden for writing the original code. It may look different now but the core logic is much the same. In my testing it mostly mitigates up to 1% packet loss between the streamer <-> ingest when packet arrival jitter is less than 30ms.